# src/main/resources/swagger30.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "Demo Notes API",
    "contact" : {
      "email" : "olivier3t@gmail.com"
    }
  },
  "paths" : {
    "/notes" : {
      "get" : {
        "summary" : "List all notes.",
        "operationId": "getNotes",
        "parameters": [],
        "responses": {
          "200" : {
            "description" : "List loaded successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type" : "array",
                  "items" : {
                    "$ref": "#/components/schemas/Note"
                  }
                }
              }
            }
          }
        },
        "security" : []
      },
      "post" : {
        "summary" : "Add a new note.",
        "operationId": "addNote",
        "parameters": [],
        "requestBody": {
          "content": {
              "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/Note"
                  }
              }
          },
          "required": true
        },
        "responses": {
          "201" : {
            "description" : "Note added successfully.",
            "content": {
              "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/Note"
                  }
              }
            }
          },
          "405": {
              "description": "Invalid input"
          }
        },
        "security" : []
      },
      "put" : {
        "summary" : "Update a note.",
        "operationId": "updateNote",
        "parameters": [],
        "requestBody": {
          "content": {
              "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/Note"
                  }
              }
          },
          "required": true
        },
        "responses": {
          "200" : {
            "description" : "Note updated successfully.",
            "content": {
              "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/Note"
                  }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Note not found"
          },
          "405": {
              "description": "Invalid input"
          }
        },
        "security" : []
      }
    },
    "/notes/{noteId}": {
      "get": {
          "summary": "Find note by ID",
          "operationId": "getNoteById",
          "parameters": [
              {
                  "name": "noteId",
                  "in": "path",
                  "description": "ID of note to return",
                  "required": true,
                  "schema": {
                      "type": "integer",
                      "format": "int64"
                  }
              }
          ],
          "responses": {
              "200": {
                  "description": "Note successfully retrieved",
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Note"
                          }
                      }
                  }
              },
              "400": {
                  "description": "Invalid ID supplied"
              },
              "404": {
                  "description": "Note not found"
              }
          },
          "security": []
      },
      "delete": {
          "summary": "Deletes a note",
          "operationId": "deleteNote",
          "parameters": [
              {
                  "name": "noteId",
                  "in": "path",
                  "description": "Note id to delete",
                  "required": true,
                  "schema": {
                      "type": "integer",
                      "format": "int64"
                  }
              }
          ],
          "responses": {
              "200": {
                "description": "Note successfully deleted"
              },
              "400": {
                  "description": "Invalid ID supplied"
              },
              "404": {
                  "description": "Note not found"
              }
          },
          "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "Note" : {
        "type": "object",
        "properties": {
            "id": {
                "type": "integer",
                "format": "int64"
            },
            "subject": {
                "type": "string"
            },
            "body": {
              "type": "string"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "date": {
                "type": "string",
                "format": "date-time"
            }
        }
      },
      "ApiMessage" : {
        "type" : "object",
        "required" : [ "code" ],
        "properties" : {
          "code" : {
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          }
        }
      }  
    }
  }
}